pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/Timur-Latypov/python-project-50"
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3.1",
    "Operating System :: POSIX :: Linux"
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

